Remove unused textbox imports
When we converted textboxes on Notify to govuk-frontend design system
style with @tombye , we did not remove the imports.
This work is now done in this commit.

Some things that will still stay in our repo for now:

textbox macro in our components folder and associated css
textboxes which are really textarea fields (🤷 how do they work)
textboxes on the styleguide page

We should revisit this when we convert textarea fields
to govuk-frontend.