Remove issue with CSV table for low vision users
The following issues were raised with the table
that replays CSV data to users:
1. the table could not be located by low vision
   users using the reflow technique
2. the content should be presented in a single
   column when the reflow technique is used

Number 2. came from the Web Content Accessibility
guidelines (WCAG) success criteria 1.4.10 Reflow.

I wasn't able to reproduce number 1. so asked the
Digital Accessibility Centre (DAC), who tested it
originally, for help. Tom Shaw from DAC kindly
retested it and found the problem was gone so I am
considering it fixed.

I am treating number 2. as a misinterpretation of
the success criteria as it lists data tables as an
exception to the rule:

"Except for parts of the content which require
two-dimensional layout for usage or meaning.

...Examples of content which requires
two-dimensional layout are images, maps, diagrams,
video, games, presentations, data tables, and
interfaces where it is necessary to keep toolbars
in view while manipulating content."

The full page is here:

https://www.w3.org/WAI/WCAG21/Understanding/reflow.html