Remove hack for bug in Chrome autocomplete
We added an extra, hidden, <input> to our /sign-in
and /register forms to stop Chrome's form
heuristics filling the fields in wrong.

See the commit that added it:

https://github.com/alphagov/notifications-admin/commit/33b15cdec6ce4eb4ad70a5e0701a2e49466cf825

This removes it after testing with the following
Chrome/OS combinations and all working without the
hack:

- Chrome 87, Windows 7
- Chrome 86, Windows 7
- Chrome 86, Windows 10
- Chrome 85, Windows 7
- Chrome 80, Windows 10
- Chrome 81, Windows 7
- Chrome 52, Windows 7
- Chrome 68, Windows 7
- Chrome 78, Windows 8.1
- Chrome 86, Windows 8.1

These combinations were based on the most-used
versions recorded in our analytics for the last 3
months.