Add service history for platform admins
We have a service history page that exists but people may not know
about. The only way to get to it is by knowing the URL as we don't
appear to link to it anywhere in the admin app.

This adds an easy link just for platform admins as this page may be of
some use I believe. There is a chance that the service history is not
super up to date and could see further improvements but it might still
be helpful when doing support type things.