Make fixtures more DRY
Moves a commonly used fixture into
`mock_get_service_settings_page_common` so we don't need to keep writing
it every time.

Note, we may be able to do similar in the future with

- single_reply_to_email_address
- single_letter_contact_block
- single_sms_sender

but need a bit more thought about fixing tests that would fail due to
this change and need tweaks to the order of their arguments.