Remove 2020/21 free allowance data
🚨 Do not merge until after 1 April 2020 🚨

Once this date has past we no longer need to give any services the
previous allowances, so we can remove them from the codebase to avoid
confusion.

It’s possible we change the allowance structure again, but it might
change in a way that this config-based logic doesn’t account for (what
if we did a per-organisation allowance for example). Having both years’
allowances in the config was a quick fix, not a foundation to build on.