Don't show user as unknown for service history
A commit was added:
https://github.com/alphagov/notifications-admin/commit/600e3affc12947d5a7b957b71e3d2864467d6bc2

In it, it falls back to the string 'Unknown' for actions done by those
not belonging to the service.

This commit changes the behaviour such that if the user is not in the
list of active users for a service, it will go get the user from the DB
(or redis). This should be fine to do as redis will protect us from most
calls as most of these cases are for platform admins.

This will mean we can now see which user platform admin put a service
live rather than seeing 'Unknown'.