Fix comments in forms/fields/radios/template.njk
The template is a duplicate of the GOVUK Frontend,
with some fixes to:
- let it do a few extra things we need
- let it work when converted to Jinja by the
  govuk_frontend_jinja extension

This also includes the changing back of
params['items'] to params.items. This was
originally changed to prevent a collision with the
`.items` property on dictionaries but the
extension now handles that.

The original comments were mentioned as a bit
confusing in a review of the PR for this work:

https://github.com/alphagov/notifications-admin/pull/3792#discussion_r613998489