Explain use of render_as_list property
It's not clear what this does on nested fields as
they render in 2 stages:
1. all descendant items
2. the top-level items (with HTML from descendant
   items sent in)

...therefore it's not clear setting it only
effects stage 2.

This also adds render_as_list to the base
GovukRadiosField class to match
GovukCheckboxesField and make sure the default is
false. This could potentially change some field
instances so they don't render as lists but this
is the desired behaviour unless explicitly set,
which would override the value on the base class.

This is to clear up confusion mentioned in this
comment on the related PR:

https://github.com/alphagov/notifications-admin/pull/3792#discussion_r614002957