Remove some of the option we have for broadcast services.
This is the first step in getting to an easier UI for setting the
options for a broadcast service. Here we remove the options for "live"
test channels. And change "All networks" to the "test" channel rather
than "severe".

This PR could go out now. However, there iss further improvement to
come.