Rename "app_" fixture to just "app"
This naming was introduced in 2016 without explanation [1]. I find it
confusing because:

- It's reminiscent of "_app", which is a Python convention indicating
the variable is internal, so maybe avoid using it.

- It suggests there's some other "app" fixture I should be using (there
isn't, though).

Our other repos use the app name (e.g. "notify_api") or just "app" for
the equivalent fixture. I'm not aware suffixing an underscore like this
has any meaning in Python, or in this codebase.

Let's fix the weird naming!

[1]: https://github.com/alphagov/notifications-admin/commit/3b1d521c10a515dab85700a1103912cf47ed414c